Rearragned content of views.py more and refactored accountInformation
